CC/Flask Api/requirements.txt
Flask
gunicorn
tensorflow
numpy
pillow
matplotlib
keras